+++ src/types/index.ts
@@ -3,3 +3,4 @@ export * from "./tools.js";
 export * from "./streams.js";
 export * from "./providers.js";
 export * from "./models.js";
+export * from "./multi-modal.js";